+++ chimeric-lib/fabric/src/main/java/com/chimericdream/lib/fabric/client/ChimericLibFabricClient.java
@@ -1,10 +1,12 @@
 package com.chimericdream.lib.fabric.client;
 
+import com.chimericdream.lib.fabric.trims.TrimmedArmorModelLoadingPlugin;
 import net.fabricmc.api.ClientModInitializer;
 
 public final class ChimericLibFabricClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         // This entrypoint is suitable for setting up client-specific logic, such as rendering.
+        TrimmedArmorModelLoadingPlugin.register();
     }
 }